+++ CHANGELOG.md
@@ -12,5 +12,11 @@
 
 * Added Per-Material FlatShading Option
 
+## \[0.4.0] - 7/10/2026
+
+* Revised MaterialDatabase to include fallback Colors/Sliders, and alternate textures/fallbacks.
+* Revised VoxelData to contain a bit for if the voxel is "mature" or not.
+* Added Shader Capability to render alternate textures/fallbacks as moss or snow or discoloration, but only for mature voxels.
+